fix: clear clippy -D warnings so Core CI can pass Format
Derive Default for VerifyEvidence/DriveKind, drop redundant closures,
and apply the rustc 1.96 clippy lints that blocked the 0.2 core job.

+++ crates/hi-tui/src/btw.rs
@@ -228,7 +228,7 @@ pub(crate) fn render_btw_panel(
     };
     match state {
         BtwOverlayState::Loading { .. } => {
-            let spinner = SPINNER[(tick as usize) % SPINNER.len()];
+            let spinner = SPINNER[tick % SPINNER.len()];
             let style = Style::default().fg(theme.gray).bg(bg);
             frame.render_widget(
                 Paragraph::new(Line::from(vec![
